python/requirements.txt
pyswisseph>=2.10.3
requests>=2.31.0
python-dateutil>=2.8.2
rich>=13.7.0
geopy>=2.4.0
pytz>=2024.1
ollama>=0.1.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
